LibJsonValidator: Added output of path if id of a node is not given
For that, a pointer to parent node must be existent. This has been
added to the JsonSchemaNode and to the respective constructors.

The raw pointer should be ok, as children only live as long as they
are contained in the parent node.

The jsonvalidator executable now handles the return values of the
parser better (correctly) and the root node is not moved out by default.