LibJsonValidator: Basic support for multiple types (type property is an array)
This patchset adds the possibility to have multiple types defined in the types
property. This is useful to have e.g. not conditions where the element shall
not be of some specific types. Or also to allow some values to be of different
kinds.

Even I'm not sure if this is the full solution, the tests of the test suite pass :-)