LibJsonValidator: Updated json_pointer calculation, improved validation errors
Some errors are "removed", as the output is not relevant for the validation of
the node. E.g. when validating the anyOf property of a node, the result of the
failed subnodes would not be of interest. But the errors that non of the anyOf
is valid is important. This has been fixed in general in this patch.